refactor(repl): remove old repl
