+++ internal/repl/handlers.go
@@ -1,4 +1,4 @@
-package normal2
+package repl
 
 import (
 	"log/slog"
